# specs/001-game-management/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Two Truths and a Lie Game Management API
  description: REST API for managing game sessions, teams, episodes, voting, and scoring
  version: 1.0.0
  contact:
    name: Game Management Team
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://usohonto.example.com/api
    description: Production
paths:
...
  /episodes:
    post:
      summary: Register episodes
      description: Submit 3 episodes (2 truths, 1 lie) for a participant
      operationId: registerEpisodes
      tags:
        - Episodes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterEpisodesRequest'
      responses:
        '201':
          description: Episodes registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  participantId:
                    type: string
                    format: uuid
                  episodes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Episode'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update episodes
      description: Edit episodes before team's turn starts
      operationId: updateEpisodes
      tags:
        - Episodes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterEpisodesRequest'
      responses:
        '200':
          description: Episodes updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  participantId:
                    type: string
                    format: uuid
                  episodes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Episode'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: Cannot edit episodes during or after team's turn
        '404':
          $ref: '#/components/responses/NotFound'
...
  /turns/{turnId}/reveal:
    post:
      summary: Reveal answer
      description: Show correct answer, calculate scores, and update scoreboard (host only)
      operationId: revealAnswer
      tags:
        - Turns
      parameters:
        - name: turnId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Answer revealed and scores calculated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurnRevealResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
...
  schemas:
...
    Episode:
      type: object
      required:
        - id
        - episodeNumber
        - text
      properties:
        id:
          type: string
          format: uuid
        episodeNumber:
          type: integer
          minimum: 1
          maximum: 3
        text:
          type: string
          minLength: 10
          maxLength: 500
          example: "I once climbed Mount Fuji"
        # Note: isLie field never exposed to clients
...
    TurnRevealResponse:
      type: object
      required:
        - correctEpisodeNumber
        - votes
        - pointsAwarded
        - updatedScores
      properties:
        correctEpisodeNumber:
          type: integer
          minimum: 1
          maximum: 3
        votes:
          type: array
          items:
            type: object
            properties:
              teamId:
                type: string
                format: uuid
              teamName:
                type: string
              selectedEpisodeNumber:
                type: integer
              isCorrect:
                type: boolean
        pointsAwarded:
          type: object
          properties:
            presentingTeamPoints:
              type: integer
            correctGuessingTeams:
              type: array
              items:
                type: object
                properties:
                  teamId:
                    type: string
                    format: uuid
                  points:
                    type: integer
        updatedScores:
          type: object
          additionalProperties:
            type: integer
          description: Map of teamId to new cumulative score
    RegisterEpisodesRequest:
      type: object
      required:
        - participantId
        - episodes
      properties:
        participantId:
          type: string
          format: uuid
        episodes:
          type: array
          minItems: 3
          maxItems: 3
          items:
            type: object
            required:
              - episodeNumber
              - text
              - isLie
            properties:
              episodeNumber:
                type: integer
                minimum: 1
                maximum: 3
              text:
                type: string
                minLength: 10
                maxLength: 500
              isLie:
                type: boolean
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error context
          additionalProperties: true
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Operation not allowed for this user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: header
      name: X-Session-ID
      description: Session ID for participant authentication
security:
  - SessionAuth: []
...